Finished Q1
I'm having issues wiht Q1 i h and Q 1 ii a as when i try to riun the code it says the function question1_i_h() is not defined and when i chekced the .py file i see that it is not defined so I made a function named that and returned 4 (not sure if it's right) but it still does the saem error. Also for question 1 ii a, this time the function is defined in .py but still gives me a nameerror.